Force YDB_NOTTP constant to be uint64_t type (instead of default int) as it is only needed in SimpleThreadAPI functions where a uint64_t type is expected

+++ sr_unix/libyottadb.h
@@ -83,7 +83,7 @@ enum
 #endif
 
 /* Value for tptoken for SimpleThreadAPI calls if NOT in a TP transaction */
-#define	YDB_NOTTP	0
+#define	YDB_NOTTP	((uint64_t)0)
 
 /* Macro to create/fill-in a ydb_buffer_t structure from a C string literal (string constant) */
 #define	YDB_LITERAL_TO_BUFFER(LITERAL, BUFFERP)					\
